:recycle: Move for internal functions
