+++ generator/constructor.go
@@ -2,7 +2,6 @@ package generator
 
 import (
 	"html/template"
-	"path/filepath"
 
 	"github.com/constructor/structure"
 )
@@ -42,19 +41,4 @@ func (impl Constructor) Generate() {
 	for _, component := range generateSources {
 		impl.FileWriter.Write(component.DestinationPath, component.Content())
 	}
-}
-func definitions(yaml structure.Yaml) []structure.Definition {
-	return yaml.Definitions
-}
-
-func templateFilePaths(definition structure.Definition) []structure.Path {
-	return definition.TemplateFilePaths
-}
-
-func sourceFilePaths(definition structure.Definition) []structure.Path {
-	filePaths, err := filepath.Glob(definition.SourcePath)
-	if err != nil {
-		panic(err)
-	}
-	return filePaths
 }
